requirements.txt
flask==2.3.3
flask-sock==0.6.0
twilio==8.10.0
groq==0.4.2
webrtcvad==2.0.10
numpy==1.24.3
python-dotenv==1.0.0